huge commit related to #166, testing also adding enable/disable to modulations size-optimized serialization code itself, mostly unfolding templates size-optimization: some enum code, CCASSERT string processing size-optimized json output,  - shortening key names and enum values  - key names are no longer quoted, saving some 15kb. added code for comparing AppSettings for test purposes added code for analyzing JSON space consumption stats test project now uses actual port from arduino string / print / et al

+++ src/clarinoid/application/Display.hpp
@@ -514,13 +514,17 @@ struct _CCDisplay : IDisplay
         }
         if (angleSweep >= 0)
         {
-            drawLine(origin.x, origin.y, origin.x + pd.p0.x, origin.y + pd.p0.y, [&](int x, int y, bool) {
+            drawLine(int(origin.x), int(origin.y), int(origin.x + pd.p0.x), (int)(origin.y + pd.p0.y), [&](int x, int y, bool) {
                 mDisplay.drawPixel(x, y, SSD1306_WHITE);
             });
         }
         else
         {
-            drawLine(origin.x, origin.y, origin.x + pd.p1.x, origin.y + pd.p1.y, [&](int x, int y, bool) {
+            drawLine(int(origin.x),
+                     int(origin.y),
+                     int(origin.x + pd.p1.x),
+                     int(origin.y + pd.p1.y),
+                     [&](int x, int y, bool) {
                 mDisplay.drawPixel(x, y, SSD1306_WHITE);
             });
         }
